Split analyzer tests
