+++ tests/Immediate.Validations.Tests/CodeFixTests/AddParameterToValidatePropertyMethodCodefixProviderTests.cs
@@ -1,87 +1,12 @@
-using System.Diagnostics.CodeAnalysis;
 using Immediate.Validations.Analyzers;
 using Immediate.Validations.CodeFixes;
 
 namespace Immediate.Validations.Tests.CodeFixTests;
 
-[SuppressMessage("Naming", "CA1707:Identifiers should not contain underscores", Justification = "Test methods")]
-public sealed class ValidatorClassCodefixTests
+public sealed class AddParameterToValidatePropertyMethodCodefixProviderTests
 {
 	[Fact]
-	public async Task AddValidateMethod() =>
-		await CodeFixTestHelper.CreateCodeFixTest<ValidatorClassAnalyzer, AddValidateMethodCodefixProvider>(
-			$$"""
-			namespace Immediate.Validations.Shared;
-			
-			public sealed class {|IV0001:TestAttribute|} : ValidatorAttribute
-			{
-				public const string DefaultMessage = "";
-			}
-			""",
-			$$"""
-			namespace Immediate.Validations.Shared;
-
-			public sealed class TestAttribute : ValidatorAttribute
-			{
-				public static bool ValidateProperty<T>(T value) => default;
-			
-				public const string DefaultMessage = "";
-			}
-			"""
-		).RunAsync();
-
-	[Fact]
-	public async Task MakeValidatePropertyMethodStatic() =>
-		await CodeFixTestHelper.CreateCodeFixTest<ValidatorClassAnalyzer, MakeValidatePropertyMethodStaticCodefixProvider>(
-			$$"""
-			namespace Immediate.Validations.Shared;
-			
-			public sealed class TestAttribute : ValidatorAttribute
-			{
-				public bool {|IV0002:ValidateProperty|}<T>(T value) => default;
-			
-				public const string DefaultMessage = "";
-			}
-			""",
-			$$"""
-			namespace Immediate.Validations.Shared;
-
-			public sealed class TestAttribute : ValidatorAttribute
-			{
-				public static bool ValidateProperty<T>(T value) => default;
-			
-				public const string DefaultMessage = "";
-			}
-			"""
-		).RunAsync();
-
-	[Fact]
-	public async Task CorrectValidatePropertyReturnType() =>
-		await CodeFixTestHelper.CreateCodeFixTest<ValidatorClassAnalyzer, CorrectValidatePropertyReturnTypeCodefixProvider>(
-			$$"""
-			namespace Immediate.Validations.Shared;
-			
-			public sealed class TestAttribute : ValidatorAttribute
-			{
-				public static string {|IV0004:ValidateProperty|}<T>(T value) => default;
-			
-				public const string DefaultMessage = "";
-			}
-			""",
-			$$"""
-			namespace Immediate.Validations.Shared;
-
-			public sealed class TestAttribute : ValidatorAttribute
-			{
-				public static bool ValidateProperty<T>(T value) => default;
-			
-				public const string DefaultMessage = "";
-			}
-			"""
-		).RunAsync();
-
-	[Fact]
-	public async Task AddParameterToValidatePropertyMethod_FromProperty() =>
+	public async Task AddParameterFromProperty() =>
 		await CodeFixTestHelper.CreateCodeFixTest<ValidatorClassAnalyzer, AddParameterToValidatePropertyMethodCodefixProvider>(
 			$$"""
 			using Immediate.Validations.Shared;
@@ -112,7 +37,7 @@ public sealed class ValidatorClassCodefixTests
 		).RunAsync();
 
 	[Fact]
-	public async Task AddParameterToValidatePropertyMethod_FromPrimaryConstructorParameter() =>
+	public async Task AddParameterFromPrimaryConstructorParameter() =>
 		await CodeFixTestHelper.CreateCodeFixTest<ValidatorClassAnalyzer, AddParameterToValidatePropertyMethodCodefixProvider>(
 			$$"""
 			using Immediate.Validations.Shared;
@@ -139,7 +64,7 @@ public sealed class ValidatorClassCodefixTests
 		).RunAsync();
 
 	[Fact]
-	public async Task AddParameterToValidatePropertyMethod_FromConstructorParameter() =>
+	public async Task AddParameterFromConstructorParameter() =>
 		await CodeFixTestHelper.CreateCodeFixTest<ValidatorClassAnalyzer, AddParameterToValidatePropertyMethodCodefixProvider>(
 			$$"""
 			using Immediate.Validations.Shared;
@@ -178,4 +103,5 @@ public sealed class ValidatorClassCodefixTests
 			}
 			"""
 		).RunAsync();
+
 }
